completed book lending module

+++ WebContent/listbooks.jsp
@@ -1,6 +1,12 @@
-<%@ include file="admin/header.jsp" %>
+<%
+if(request.getAttribute("userLevel").equals("member")){ %>
+	<%@ include file="header.jsp" %>
+<%}else{%>
+	<%@ include file="admin/header.jsp" %>
+<%}
+%>
 <div class="content">
-	<table border=1 class="CSSTableGenerator">
+	<table border=1 >
         <thead>
             <tr>
                 <th>Book ID</th>
@@ -11,7 +17,15 @@
                 <th>Rack Number</th>
                 <th>Publisher</th>
                 <th>Edition</th>
-                <th colspan=2>Action</th>
+                                    <%
+if(request.getAttribute("userLevel").equals("admin")){ %>
+	<th colspan=3>Action</th>
+<%}else{%>
+	<th>Action</th>
+<%}
+%>
+                
+                
             </tr>
         </thead>
         <tbody>
@@ -20,14 +34,21 @@
                     <td><c:out value="" /> ${book.id}</td>
                     <td><c:out value="" /> ${book.title}</td>
                     <td><c:out value="" /> ${book.isbn}</td>
-                    <td><c:out value="" /> ${book.author}</td>
+                    <td><c:out value="" /> ${book.author1}</td>
                     <td><c:out value="" /> ${book.classNo}</td>
                     <td><c:out value="" /> ${book.rackNo}</td>
                     <td><c:out value="" /> ${book.publisher}</td>
                     <td><c:out value="" /> ${book.edition}</td>
-                    
-                    <td><a href="BookController?action=edit&bookId=<c:out value="${book.id}"/>">Update</a></td>
+                    <%
+if(request.getAttribute("userLevel").equals("admin")){ %>
+					<td><a href="BookController?action=edit&bookId=<c:out value="${book.id}"/>">Update</a></td>
                     <td><a href="BookController?action=delete&bookId=<c:out value="${book.id}"/>">Delete</a></td>
+                    <td><a href="BookController?action=viewbook&bookId=<c:out value="${book.id}"/>">View Details</a></td>
+<%}else{%>
+	 <td><a href="BookController?action=viewbook&bookId=<c:out value="${book.id}"/>">View Details</a></td>
+<%}
+%>
+                    
                 </tr>
             </c:forEach>
         </tbody>
